Moved Unicorn to separate node module
